+++ backend/src/main/java/org/iclass/backend/dto/BookMarkDto.java
@@ -1,8 +1,8 @@
 package org.iclass.backend.dto;
 
-import org.iclass.backend.asdf.BookMarkEntity;
-import org.iclass.backend.asdf.MovieInfoEntity;
-import org.iclass.backend.asdf.UsersEntity;
+import org.iclass.backend.entity.BookMarkEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
+import org.iclass.backend.entity.UsersEntity;
 
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
