+++ backend/src/main/java/org/iclass/backend/dto/MovieCastDto.java
@@ -1,6 +1,6 @@
 package org.iclass.backend.dto;
 
-import org.iclass.backend.asdf.MovieCastEntity;
+import org.iclass.backend.entity.MovieCastEntity;
 
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
