+++ backend/src/main/java/org/iclass/backend/dto/MovieGenresDto.java
@@ -1,8 +1,8 @@
 package org.iclass.backend.dto;
 
-import org.iclass.backend.asdf.GenresEntity;
-import org.iclass.backend.asdf.MovieGenresEntity;
-import org.iclass.backend.asdf.MovieInfoEntity;
+import org.iclass.backend.entity.GenresEntity;
+import org.iclass.backend.entity.MovieGenresEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
 
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
