+++ backend/src/main/java/org/iclass/backend/dto/MovieInfoDto.java
@@ -2,7 +2,7 @@ package org.iclass.backend.dto;
 
 import java.time.LocalDate;
 
-import org.iclass.backend.asdf.MovieInfoEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
 
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
