+++ backend/src/main/java/org/iclass/backend/dto/SoundTrackDto.java
@@ -1,7 +1,7 @@
 package org.iclass.backend.dto;
 
-import org.iclass.backend.asdf.MovieInfoEntity;
-import org.iclass.backend.asdf.SoundTrackEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
+import org.iclass.backend.entity.SoundTrackEntity;
 
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
