+++ backend/src/main/java/org/iclass/backend/dto/UsersDto.java
@@ -2,7 +2,7 @@ package org.iclass.backend.dto;
 
 import java.time.LocalDateTime;
 
-import org.iclass.backend.asdf.UsersEntity;
+import org.iclass.backend.entity.UsersEntity;
 
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
