+++ backend/src/main/java/org/iclass/backend/entity/GenresEntity.java
@@ -1,12 +1,10 @@
-package org.iclass.backend.asdf;
+package org.iclass.backend.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -17,24 +15,22 @@ import lombok.ToString;
 
 @Getter
 @Setter
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @ToString
 @Builder
 @Entity
-@Table(name = "MOVIE_GENRES")
-public class MovieGenresEntity {
+@Table(name = "GENRES")
+public class GenresEntity {
 
   @Id
-  @Column(name = "MG_idx")
+  @Column(name = "genre_idx")
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long mgIdx;
+  private Long genreIdx;
 
-  @ManyToOne
-  @JoinColumn(name = "movie_idx", nullable = false)
-  private MovieInfoEntity movie;
+  @Column(name = "genre_id", nullable = false)
+  private Long genreId;
 
-  @ManyToOne
-  @JoinColumn(name = "genre_id", referencedColumnName = "genre_idx", nullable = false)
-  private GenresEntity genre;
+  @Column(nullable = false, length = 50)
+  private String name;
 }
