+++ backend/src/main/java/org/iclass/backend/entity/MovieCrewEntity.java
@@ -1,4 +1,4 @@
-package org.iclass.backend.asdf;
+package org.iclass.backend.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,34 +13,31 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@Table(name = "MOVIE_CAST")
+@Table(name = "MOVIE_CREW")
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class MovieCastEntity {
+public class MovieCrewEntity {
 
     @Id
-    @Column(name = "CAST_IDX")
+    @Column(name = "CREDIT_IDX")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long castIdx;
+    private Long creditIdx;
 
     @Column(name = "TMDB_MOVIE_ID", nullable = false)
     private Long tmdbMovieId;
 
-    @Column(name = "TMDB_CAST_ID", nullable = false)
-    private Long tmdbCastId;
+    @Column(name = "TMDB_CREW_ID")
+    private Long tmdbCrewId;
 
-    @Column(name = "CHARACTER", length = 255)
-    private String character;
+    @Column(name = "CREW_NAME", length = 255)
+    private String crewName;
 
-    @Column(name = "CAST_NAME", length = 255)
-    private String castName;
+    @Column(name = "CREW_PROFILE_PATH", length = 255)
+    private String crewProfilePath;
 
-    @Column(name = "CAST_PROFILE_PATH", length = 255)
-    private String castProfilePath;
-
-    @Column(name = "CREDIT_ORDER")
-    private Integer creditOrder;
+    @Column(name = "JOB", length = 255)
+    private String job;
 }
