+++ backend/src/main/java/org/iclass/backend/entity/MovieInfoEntity.java
@@ -1,4 +1,4 @@
-package org.iclass.backend.asdf;
+package org.iclass.backend.entity;
 
 import java.time.LocalDate;
