+++ backend/src/main/java/org/iclass/backend/entity/MovieVsEntity.java
@@ -1,10 +1,12 @@
-package org.iclass.backend.asdf;
+package org.iclass.backend.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,22 +17,28 @@ import lombok.ToString;
 
 @Getter
 @Setter
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @ToString
 @Builder
 @Entity
-@Table(name = "GENRES")
-public class GenresEntity {
+@Table(name = "MOVIE_VS")
+public class MovieVsEntity {
 
   @Id
-  @Column(name = "genre_idx")
+  @Column(name = "VS_idx")
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long genreIdx;
+  private Long vsIdx;
 
-  @Column(name = "genre_id", nullable = false)
-  private Long genreId;
+  @ManyToOne
+  @JoinColumn(name = "movie_VS1", nullable = false)
+  private MovieInfoEntity movieVs1;
 
-  @Column(nullable = false, length = 50)
-  private String name;
+  @ManyToOne
+  @JoinColumn(name = "movie_VS2", nullable = false)
+  private MovieInfoEntity movieVs2;
+
+  @Column
+  @Builder.Default
+  private Integer active = 0;
 }
