+++ backend/src/main/java/org/iclass/backend/repository/ArticlesRepository.java
@@ -2,8 +2,8 @@ package org.iclass.backend.repository;
 
 import java.util.List;
 
-import org.iclass.backend.asdf.ArticlesEntity;
-import org.iclass.backend.asdf.MovieInfoEntity;
+import org.iclass.backend.entity.ArticlesEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
