+++ backend/src/main/java/org/iclass/backend/repository/CommentsRepository.java
@@ -2,8 +2,8 @@ package org.iclass.backend.repository;
 
 import java.util.List;
 
-import org.iclass.backend.asdf.CommentsEntity;
-import org.iclass.backend.asdf.ReviewEntity;
+import org.iclass.backend.entity.CommentsEntity;
+import org.iclass.backend.entity.ReviewEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
