+++ backend/src/main/java/org/iclass/backend/repository/MovieCastRepository.java
@@ -2,7 +2,7 @@ package org.iclass.backend.repository;
 
 import java.util.List;
 
-import org.iclass.backend.asdf.MovieCastEntity;
+import org.iclass.backend.entity.MovieCastEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
