+++ backend/src/main/java/org/iclass/backend/repository/MovieCrewRepository.java
@@ -2,7 +2,7 @@ package org.iclass.backend.repository;
 
 import java.util.List;
 
-import org.iclass.backend.asdf.MovieCrewEntity;
+import org.iclass.backend.entity.MovieCrewEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
