+++ backend/src/main/java/org/iclass/backend/repository/MovieGenresRepository.java
@@ -1,8 +1,8 @@
 package org.iclass.backend.repository;
 
-import org.iclass.backend.asdf.GenresEntity;
-import org.iclass.backend.asdf.MovieGenresEntity;
-import org.iclass.backend.asdf.MovieInfoEntity;
+import org.iclass.backend.entity.GenresEntity;
+import org.iclass.backend.entity.MovieGenresEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
