+++ backend/src/main/java/org/iclass/backend/repository/MovieInfoRepository.java
@@ -2,7 +2,7 @@ package org.iclass.backend.repository;
 
 import java.util.Optional;
 
-import org.iclass.backend.asdf.MovieInfoEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
