+++ backend/src/main/java/org/iclass/backend/repository/MovieVSRepository.java
@@ -1,6 +1,6 @@
 package org.iclass.backend.repository;
 
-import org.iclass.backend.asdf.MovieVsEntity;
+import org.iclass.backend.entity.MovieVsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
