+++ backend/src/main/java/org/iclass/backend/repository/MovieVoteRepository.java
@@ -2,9 +2,9 @@ package org.iclass.backend.repository;
 
 import java.util.Optional;
 
-import org.iclass.backend.asdf.MovieVoteEntity;
-import org.iclass.backend.asdf.MovieVsEntity;
-import org.iclass.backend.asdf.UsersEntity;
+import org.iclass.backend.entity.MovieVoteEntity;
+import org.iclass.backend.entity.MovieVsEntity;
+import org.iclass.backend.entity.UsersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
