+++ backend/src/main/java/org/iclass/backend/repository/NoticeRepository.java
@@ -1,6 +1,6 @@
 package org.iclass.backend.repository;
 
-import org.iclass.backend.asdf.NoticeEntity;
+import org.iclass.backend.entity.NoticeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
