+++ backend/src/main/java/org/iclass/backend/repository/RankingRepository.java
@@ -2,8 +2,8 @@ package org.iclass.backend.repository;
 
 import java.util.List;
 
-import org.iclass.backend.asdf.MovieInfoEntity;
-import org.iclass.backend.asdf.RankingEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
+import org.iclass.backend.entity.RankingEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
