+++ backend/src/main/java/org/iclass/backend/repository/SoundTrackRepository.java
@@ -2,7 +2,7 @@ package org.iclass.backend.repository;
 
 import java.util.List;
 
-import org.iclass.backend.asdf.SoundTrackEntity;
+import org.iclass.backend.entity.SoundTrackEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
