+++ backend/src/main/java/org/iclass/backend/repository/UsersRepository.java
@@ -2,7 +2,7 @@ package org.iclass.backend.repository;
 
 import java.util.Optional;
 
-import org.iclass.backend.asdf.UsersEntity;
+import org.iclass.backend.entity.UsersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
