+++ backend/src/main/java/org/iclass/backend/repository/VideosRepository.java
@@ -2,8 +2,8 @@ package org.iclass.backend.repository;
 
 import java.util.List;
 
-import org.iclass.backend.asdf.MovieInfoEntity;
-import org.iclass.backend.asdf.VideosEntity;
+import org.iclass.backend.entity.MovieInfoEntity;
+import org.iclass.backend.entity.VideosEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
