+++ backend/src/main/java/org/iclass/backend/security/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package org.iclass.backend.security;
 
 import lombok.RequiredArgsConstructor;
 
-import org.iclass.backend.asdf.UsersEntity;
+import org.iclass.backend.entity.UsersEntity;
 import org.iclass.backend.repository.UsersRepository;
 import org.springframework.security.core.userdetails.*;
 import org.springframework.stereotype.Service;
